Factor out etcdWatcher to a separate file
